requirements.txt
boto3==1.35.45
botocore==1.35.45
dateutils==0.6.12
jmespath==1.0.1
python-dateutil==2.9.0.post0
pytz==2024.2
s3transfer==0.10.3
shortuuid==1.0.13
six==1.16.0
tqdm==4.66.5
urllib3==1.26.20
